+++ src/resolver/setting/setting.resolver.ts
@@ -1,20 +1,14 @@
 import {Injectable} from '@nestjs/common'
-import {ConfigService} from '@nestjs/config'
 import {Args, ID, Query} from '@nestjs/graphql'
 import {Roles} from '../../decorator/roles.decorator'
 import {User} from '../../decorator/user.decorator'
 import {PagerSettingModel} from '../../dto/setting/pager.setting.model'
 import {SettingModel} from '../../dto/setting/setting.model'
-import {UserModel} from '../../dto/user/user.model'
 import {UserDocument} from '../../schema/user.schema'
 import {SettingService} from '../../service/setting.service'
 
 @Injectable()
 export class SettingResolver {
-  private publicKeys: string[] = [
-    'SIGNUP_DISABLED',
-  ]
-
   constructor(
     private readonly settingService: SettingService,
   ) {
@@ -37,7 +31,7 @@ export class SettingResolver {
     @Args('key', {type: () => ID}) key: string,
     @User() user: UserDocument,
   ): Promise<SettingModel> {
-    if (!this.publicKeys.includes(key) && !user.roles.includes('superuser')) {
+    if (!this.settingService.isPublicKey(key) && !user.roles.includes('superuser')) {
       throw new Error(`no access to key ${key}`)
     }